Use uint64 for memory size in Domain.SetMaxMemory
This matches the type used for the other memory-related functions, as
well as the type of the C function.

+++ domain.go
@@ -1453,7 +1453,7 @@ func (d *Domain) UndefineFlags(flags DomainUndefineFlagsValues) error {
 }
 
 // See also https://libvirt.org/html/libvirt-libvirt-domain.html#virDomainSetMaxMemory
-func (d *Domain) SetMaxMemory(memory uint) error {
+func (d *Domain) SetMaxMemory(memory uint64) error {
 	var err C.virError
 	result := C.virDomainSetMaxMemoryWrapper(d.ptr, C.ulong(memory), &err)
 	if result == -1 {
